+++ app/Http/Requests/Lease/MoveLeaseRequest.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Http\Requests\Lease;
+
+use Illuminate\Contracts\Validation\ValidationRule;
+use Illuminate\Foundation\Http\FormRequest;
+
+class MoveLeaseRequest extends FormRequest
+{
+    /**
+     * @return array<string, ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'target_room_id' => ['required', 'integer', 'exists:rooms,id'],
+        ];
+    }
+}
