+++ app/Http/Requests/Lease/MoveOutRequest.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace App\Http\Requests\Lease;
+
+use Illuminate\Contracts\Validation\ValidationRule;
+use Illuminate\Foundation\Http\FormRequest;
+
+class MoveOutRequest extends FormRequest
+{
+    /**
+     * @return array<string, ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'move_out_date' => ['required', 'date'],
+            'reason' => ['nullable', 'string', 'max:255'],
+            'deposit_returned' => ['nullable', 'boolean'],
+            'deposit_refund_amount' => ['nullable', 'numeric', 'min:0'],
+            'notes' => ['nullable', 'string', 'max:65535'],
+            'move_to_another_room' => ['nullable', 'boolean'],
+            'target_room_id' => ['nullable', 'integer', 'exists:rooms,id'],
+        ];
+    }
+}
