+++ app/Http/Requests/Role/UpdateRoleRequest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Role;
 
 use App\Enums\Permission;
 use Illuminate\Contracts\Validation\ValidationRule;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
 
-class StoreRoleRequest extends FormRequest
+class UpdateRoleRequest extends FormRequest
 {
     public function authorize(): bool
     {
@@ -19,13 +19,21 @@ class StoreRoleRequest extends FormRequest
      */
     public function rules(): array
     {
-        return [
-            'name' => ['required', 'string', 'max:255', 'alpha_dash', 'unique:roles,name'],
-            'label' => ['required', 'string', 'max:255'],
+        $role = $this->route('role');
+
+        $rules = [
+            'label' => ['nullable', 'string', 'max:255'],
             'description' => ['nullable', 'string', 'max:1000'],
             'color' => ['nullable', 'string', 'regex:/^#[0-9A-Fa-f]{6}$/'],
+            'is_active' => ['boolean'],
             'permissions' => ['array'],
             'permissions.*' => ['string', Rule::in(Permission::values())],
         ];
+
+        if ($role && $role->is_system) {
+            unset($rules['is_active']);
+        }
+
+        return $rules;
     }
 }
