+++ app/Http/Requests/Tenant/AssignRoomRequest.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace App\Http\Requests\Tenant;
+
+use App\Enums\BillingUnit;
+use Illuminate\Contracts\Validation\ValidationRule;
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Validation\Rule;
+
+class AssignRoomRequest extends FormRequest
+{
+    /**
+     * @return array<string, ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'tenant_ids' => ['nullable', 'array', 'min:1'],
+            'tenant_ids.*' => ['integer', 'distinct', 'exists:tenants,id'],
+            'room_id' => ['required', 'integer', 'exists:rooms,id'],
+            'start_date' => ['required', 'date'],
+            'end_date' => ['nullable', 'date', 'after:start_date'],
+            'rent_amount' => ['nullable', 'numeric', 'min:0'],
+            'billing_interval' => ['nullable', 'integer', 'min:1', 'max:255'],
+            'billing_unit' => ['nullable', 'string', Rule::in(BillingUnit::values())],
+            'room_rate_id' => ['nullable', 'integer', 'exists:room_rates,id'],
+            'deposit_amount' => ['nullable', 'numeric', 'min:0'],
+            'deposit_paid_at' => ['nullable', 'date'],
+            'rent_due_day' => ['nullable', 'integer', 'between:1,31'],
+            'notes' => ['nullable', 'string', 'max:65535'],
+        ];
+    }
+}
