+++ app/Http/Requests/User/CompleteInvitationRequest.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace App\Http\Requests\User;
+
+use Illuminate\Contracts\Validation\ValidationRule;
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Validation\Rules\Password;
+
+class CompleteInvitationRequest extends FormRequest
+{
+    /**
+     * @return array<string, ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'token' => ['required', 'string'],
+            'email' => ['required', 'email'],
+            'password' => ['required', 'string', Password::default(), 'confirmed'],
+            'password_confirmation' => ['required', 'string'],
+        ];
+    }
+}
